refactor(agents): move cluster prompt rendering
Keep LLM-facing cluster descriptions and formatting in agents/llm_renderers rather than the cluster assembly mixin.

Amp-Thread-ID: https://ampcode.com/threads/T-01a024ad-20dd-707e-b96a-70cb73fdaa78
