+++ agents/incremental_agent.py
@@ -26,10 +26,11 @@ from agents.agent_responses import (
     iter_components,
 )
 from agents.file_index_models import FileMethodGroup, MethodEntry
-from agents.cluster_methods_mixin import ClusterMethodsMixin, cluster_group_ids, render_cluster_groups
+from agents.cluster_methods_mixin import ClusterMethodsMixin
 from agents.content_hash import SourceCache
 from agents.cluster_ids import CodeBoardingClusterIds
 from agents.incremental_results import ScopeRelationContext, ScopeUpdateResult
+from agents.llm_renderers import cluster_group_ids, render_cluster_groups
 from agents.prompts import (
     format_project_system_message,
     get_api_surfaces_message,
